refactor androidTest
