+++ app/build.gradle.kts
@@ -54,45 +54,17 @@ android {
             excludes += "/META-INF/{AL2.0,LGPL2.1}"
         }
     }
-    configurations.all {
-        resolutionStrategy {
-            force("androidx.test.ext:junit:1.1.3")
-            force("androidx.test.espresso:espresso-core:3.4.0")
-        }
-    }
 }
 
 dependencies {
-    implementation(project(":design-system"))
     implementation(project(":features:nyarticles:presentation"))
-    implementation(project(":utils"))
-    implementation("androidx.core:core-ktx:1.13.1")
+    implementation("com.google.android.material:material:1.12.0")
     implementation("androidx.lifecycle:lifecycle-runtime-ktx:2.8.0")
     implementation("androidx.activity:activity-compose:1.9.0")
     implementation(platform("androidx.compose:compose-bom:2023.08.00"))
     implementation("androidx.compose.ui:ui")
     implementation("androidx.compose.ui:ui-graphics")
-    implementation("androidx.compose.ui:ui-tooling-preview")
     implementation("androidx.compose.material3:material3")
-
     implementation("com.google.dagger:hilt-android:2.49")
-    implementation("androidx.compose.ui:ui-test-junit4-android:1.6.7")
-    testImplementation("org.junit.jupiter:junit-jupiter:5.8.2")
     kapt("com.google.dagger:hilt-compiler:2.48.1")
-    implementation("androidx.hilt:hilt-navigation-compose:1.2.0")
-
-    implementation("org.mockito.kotlin:mockito-kotlin:4.0.0")
-    implementation("org.mockito:mockito-core:4.0.0")
-    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-test:1.8.1")
-
-    testImplementation("junit:junit:4.13.2")
-    androidTestImplementation("androidx.test.ext:junit:1.1.5")
-    androidTestImplementation("androidx.test.espresso:espresso-core:3.5.1")
-    androidTestImplementation(platform("androidx.compose:compose-bom:2023.08.00"))
-    androidTestImplementation("androidx.compose.ui:ui-test-junit4")
-    debugImplementation("androidx.compose.ui:ui-tooling")
-    debugImplementation("androidx.compose.ui:ui-test-manifest")
-    androidTestImplementation("androidx.compose.ui:ui-test-junit4:1.6.7")
-    androidTestImplementation("androidx.arch.core:core-testing:2.2.0")
-    androidTestImplementation("androidx.test.espresso:espresso-core:3.5.1")
 }
